Move to new BlueZ version and switch from gatttool to bluetoothctl.
With this move it will be possible to enter existing bluetooth 
connections, making it possible to let system settings handle the 
pairing and automatic connection attempts.

To handle added devices, the app will now display paired devices. This 
also allows to display unsupported devices. The default setting is to 
only display supported devices.

The app will let the user choose if they want to display unsupported 
devices. Therefore the settings page was created and moved to 
Welcome.qml. The About.qml page was moved to the header of Settings.qml.